move locale fetch to url utils
