+++ static/js/tools/map/map.ts
@@ -17,7 +17,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import { getLocaleFromUrl, loadLocaleData } from "../../i18n/i18n";
+import { loadLocaleData } from "../../i18n/i18n";
+import { getLocaleFromUrl } from "../../utils/url_utils";
 import { AppWithContext } from "./app";
 
 window.addEventListener("load", (): void => {
